+++ src/modules/docker_build.rs
@@ -7,7 +7,7 @@ use crate::args::AppConfig;
 use crate::data::DOCKER_PACKAGES_LIST;
 use crate::data::DOCKER_TAGS_LIST;
 use crate::generators::gen_hex_string;
-use crate::io::{csleep, dprint, newline, print};
+use crate::io::{csleep, newline, print};
 use crate::modules::Module;
 use crate::ALL_MODULES;
 
@@ -31,13 +31,9 @@ impl Module for DockerBuild {
         let mut current_size: f64 = 0.0;
 
         while current_size <= target_size {
-            dprint(
-                format!(
-                    "\rSending build context to Docker daemon  {current_size:>4.2}MB",
-                    current_size = current_size
-                ),
-                0,
-            )
+            print(format!(
+                "\rSending build context to Docker daemon  {current_size:>4.2}MB",
+            ))
             .await;
 
             let remaining_size = target_size - current_size;
@@ -70,8 +66,6 @@ impl Module for DockerBuild {
             // Print the current step with the instruction to run
             print(format!(
                 "\rStep {current_step}/{total_steps} : {instruction}",
-                current_step = current_step,
-                total_steps = total_steps,
                 instruction = ["RUN", &chosen_module.signature()].join(" "),
             ))
             .await;
@@ -110,14 +104,8 @@ impl Module for DockerBuild {
         let image: &&str = DOCKER_PACKAGES_LIST.choose(&mut rng).unwrap();
         let image_tag: &&str = DOCKER_TAGS_LIST.choose(&mut rng).unwrap();
 
-        print(format!("Successfully built {hash}", hash = hash)).await;
-
-        print(format!(
-            "Successfully tagged {image}:{tag}",
-            image = image,
-            tag = image_tag
-        ))
-        .await;
+        print(format!("Successfully built {hash}")).await;
+        print(format!("Successfully tagged {image}:{image_tag}",)).await;
 
         if appconfig.should_exit() {
             return;