+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+- Add a bunch more extensions and file formats
+
+## [0.12.0] - 2022-02-21
+- Add Ansible module [#301](https://github.com/svenstaro/genact/issues/301)
 
 ## [0.11.0] - 2021-03-19
 - Add docker_build module [#103](https://github.com/svenstaro/genact/pull/103) (thanks @Kovah)
@@ -29,7 +33,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Rewrite web version using wasm-bindgen
 
 <!-- next-url -->
-[Unreleased]: https://github.com/svenstaro/genact/compare/v0.11.0...HEAD
+[Unreleased]: https://github.com/svenstaro/genact/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/svenstaro/genact/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/svenstaro/genact/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/svenstaro/genact/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/svenstaro/genact/compare/v0.8.1...v0.9.0